Fix indentation issue for leading dots
We were indenting like this:

foo
  .bar { }
.baz

because the leading-dot logic only applied if your previous token was
an identifier (not "}" ending a block). Now the leading-dot logic
applies if your dot is the first non-whitespace thing on the line
regardless of previous thing. I also tweaked the period check to not
allocate a new string each time.